dax: Fix alloc_dax_region() compile warning
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-294.el8
commit-author Dan Williams <dan.j.williams@intel.com>
commit 460370ab20b6cc174256e46e192adf01e730faf6

PFN flags are (unsigned long long), fix the alloc_dax_region() calling
convention to fix warnings of the form:

>> include/linux/pfn_t.h:18:17: warning: large integer implicitly truncated to unsigned type [-Woverflow]
    #define PFN_DEV (1ULL << (BITS_PER_LONG_LONG - 3))

	Reported-by: kbuild test robot <lkp@intel.com>
	Acked-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 460370ab20b6cc174256e46e192adf01e730faf6)

+++ drivers/dax/dax-private.h
@@ -40,7 +40,7 @@ struct dax_region {
 	struct device *dev;
 	unsigned int align;
 	struct resource res;
-	unsigned long pfn_flags;
+	unsigned long long pfn_flags;
 };
 
 /**
